refactor: address menu redesign review findings
- Unsubscribe layout subscription on destroy
- Remove dead hover-state handlers left from hover-to-expand
- Share getUserInitials helper between header and profile
- Guard no-op layout changes in MenuLayoutService
- Cache branding data instead of refetching per navigation
