+++ frontend/src/app/utils/index.ts
@@ -3,3 +3,4 @@ export { CategoryAccess, CategoryDetails, CategoryGroup } from "./permissions-ca
 export { EntityAccess, EntityGroup } from "./permissions-entity";
 export { PermissionsGroup } from "./permissions";
 export { MergeUtils } from "./merge-utils";
+export { getUserInitials } from "./user-initials";
