+++ frontend/src/app/views/root-profile/root-profile.component.ts
@@ -26,6 +26,7 @@ import { ToastrService } from 'ngx-toastr';
 import { AppTheme, AppThemeOption, AppThemeService } from '../../services/app-theme.service';
 import { MenuLayout, MenuLayoutOption, MenuLayoutService } from '../../services/menu-layout.service';
 import { DocWidgetService } from '../../services/doc-widget.service';
+import { getUserInitials } from '../../utils';
 
 enum OperationMode {
     None,
@@ -861,10 +862,7 @@ export class RootProfileComponent implements OnInit, OnDestroy {
     }
 
     getInitials(username: string | undefined): string {
-        if (!username) { return '?'; }
-        const caps = username.match(/[A-Z]/g);
-        if (caps && caps.length >= 2) { return caps.slice(0, 2).join(''); }
-        return username.slice(0, 2).toUpperCase();
+        return getUserInitials(username);
     }
 
     formatRole(role: string | undefined): string {
